ByteBuffer class to new file
